requirements.txt
# DocMind AI — requirements
# Install with: pip install -r requirements.txt

# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Streamlit UI
streamlit==1.40.2

# LangChain stack
langchain==0.3.9
langchain-community==0.3.9
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.2

# HuggingFace / local model (free, no API key needed)
transformers==4.46.3
torch==2.5.1          # CPU-only is fine; for GPU: pip install torch --index-url https://download.pytorch.org/whl/cu121
sentence-transformers==3.3.1
accelerate==1.2.1

# PDF loading
pypdf==5.1.0

# Vector store
faiss-cpu==1.9.0.post1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.2